refactor: cart Atom 명 변경(state -> cart)
