+++ src/components/cartList/CartItem.tsx
@@ -4,7 +4,7 @@ import * as S from './styles/CartItem.styles';
 import CheckBox from '../common/CheckBox';
 import QuantityInput from '../common/QuantityInput';
 import * as api from '../../api';
-import { cartState } from '../../atom/state';
+import { cartState } from '../../atom/cart';
 import { API_ERROR_MESSAGE, MAX_QUANTITY } from '../../constants';
 import useToast from '../hooks/useToast';
 import { serverNameState } from '../../atom/serverName';
